# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Catalyst Discord Bot API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OkResponse {
  success: boolean;
}

export interface LoginBody {
  password: string;
}

export interface LoginResponse {
  success: boolean;
  error?: string;
}

export interface AuthCheckResponse {
  loggedIn: boolean;
}

export interface ConnectBotBody {
  token: string;
}

export interface ConnectBotResponse {
  success: boolean;
  username?: string;
  error?: string;
}

export interface BotStatusResponse {
  connected: boolean;
  username?: string;
}

export interface Guild {
  id: string;
  name: string;
  icon?: string;
  memberCount?: number;
}

export interface GuildsResponse {
  success: boolean;
  guilds: Guild[];
}

export interface Channel {
  id: string;
  name: string;
  type: number;
}

export interface ChannelsResponse {
  success: boolean;
  channels: Channel[];
}

export interface GuildActionBody {
  guildId: string;
}

export interface ActionResult {
  success: boolean;
  name: string;
  error?: string;
}

export interface ActionResponse {
  success: boolean;
  message?: string;
  error?: string;
}

export interface BulkActionResponse {
  success: boolean;
  results?: ActionResult[];
  error?: string;
}

export interface SendMessageBody {
  guildId: string;
  channelId: string;
  message: string;
}

export type GetBotChannelsParams = {
  guildId: string;
};