# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Catalyst Discord Bot API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Login with password
 */
export const LoginBody = zod.object({
  password: zod.string(),
});

export const LoginResponse = zod.object({
  success: zod.boolean(),
  error: zod.string().optional(),
});

/**
 * @summary Logout
 */
export const LogoutResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Check authentication status
 */
export const AuthCheckResponse = zod.object({
  loggedIn: zod.boolean(),
});

/**
 * @summary Connect bot with token
 */
export const ConnectBotBody = zod.object({
  token: zod.string(),
});

export const ConnectBotResponse = zod.object({
  success: zod.boolean(),
  username: zod.string().optional(),
  error: zod.string().optional(),
});

/**
 * @summary Disconnect the bot
 */
export const DisconnectBotResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Get bot connection status
 */
export const GetBotStatusResponse = zod.object({
  connected: zod.boolean(),
  username: zod.string().optional(),
});

/**
 * @summary Get guilds the bot is in
 */
export const GetBotGuildsResponse = zod.object({
  success: zod.boolean(),
  guilds: zod.array(
    zod.object({
      id: zod.string(),
      name: zod.string(),
      icon: zod.string().optional(),
      memberCount: zod.number().optional(),
    }),
  ),
});

/**
 * @summary Get channels for a guild
 */
export const GetBotChannelsQueryParams = zod.object({
  guildId: zod.coerce.string(),
});

export const GetBotChannelsResponse = zod.object({
  success: zod.boolean(),
  channels: zod.array(
    zod.object({
      id: zod.string(),
      name: zod.string(),
      type: zod.number(),
    }),
  ),
});

/**
 * @summary Bot leaves a guild
 */
export const LeaveGuildBody = zod.object({
  guildId: zod.string(),
});

export const LeaveGuildResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
  error: zod.string().optional(),
});

/**
 * @summary Delete all channels in a guild
 */
export const DeleteChannelsBody = zod.object({
  guildId: zod.string(),
});

export const DeleteChannelsResponse = zod.object({
  success: zod.boolean(),
  results: zod
    .array(
      zod.object({
        success: zod.boolean(),
        name: zod.string(),
        error: zod.string().optional(),
      }),
    )
    .optional(),
  error: zod.string().optional(),
});

/**
 * @summary Delete all roles in a guild
 */
export const DeleteRolesBody = zod.object({
  guildId: zod.string(),
});

export const DeleteRolesResponse = zod.object({
  success: zod.boolean(),
  results: zod
    .array(
      zod.object({
        success: zod.boolean(),
        name: zod.string(),
        error: zod.string().optional(),
      }),
    )
    .optional(),
  error: zod.string().optional(),
});

/**
 * @summary Kick all members from a guild
 */
export const KickAllMembersBody = zod.object({
  guildId: zod.string(),
});

export const KickAllMembersResponse = zod.object({
  success: zod.boolean(),
  results: zod
    .array(
      zod.object({
        success: zod.boolean(),
        name: zod.string(),
        error: zod.string().optional(),
      }),
    )
    .optional(),
  error: zod.string().optional(),
});

/**
 * @summary Ban all members from a guild
 */
export const BanAllMembersBody = zod.object({
  guildId: zod.string(),
});

export const BanAllMembersResponse = zod.object({
  success: zod.boolean(),
  results: zod
    .array(
      zod.object({
        success: zod.boolean(),
        name: zod.string(),
        error: zod.string().optional(),
      }),
    )
    .optional(),
  error: zod.string().optional(),
});

/**
 * @summary Send a message to a channel
 */
export const SendMessageBody = zod.object({
  guildId: zod.string(),
  channelId: zod.string(),
  message: zod.string(),
});

export const SendMessageResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
  error: zod.string().optional(),
});