# lib/api-zod/src/generated/types/actionResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Catalyst Discord Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface ActionResult {
  success: boolean;
  name: string;
  error?: string;
}