# lib/api-zod/src/generated/types/bulkActionResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Catalyst Discord Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { ActionResult } from "./actionResult";

export interface BulkActionResponse {
  success: boolean;
  results?: ActionResult[];
  error?: string;
}