# lib/api-zod/src/generated/types/channelsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Catalyst Discord Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { Channel } from "./channel";

export interface ChannelsResponse {
  success: boolean;
  channels: Channel[];
}