# lib/api-zod/src/generated/types/guildsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Catalyst Discord Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { Guild } from "./guild";

export interface GuildsResponse {
  success: boolean;
  guilds: Guild[];
}